requirements.txt
gitpython
sentence-transformers
pinecone-client
